ZENAI-4542: Fix 32-bit overflow in LOWOHA matmul and gtest buffer sizing (#619)
* ZENAI-4542: Fix 32-bit overflow in LOWOHA matmul size and stride computations

Widen the leading operand to size_t so each product is evaluated in
64-bit instead of overflowing int before being converted.

- lru_cache/lowoha_cache.hpp: 2D zero-point compensation buffer size
  (M * N) in both the wei-zp-only and both-zp branches. The buffer was
  under-allocated on overflow while the fill loops still wrote M * N
  entries.
- lowoha_matmul_utils.cpp: partitioned post-op row offset (m_start * N),
  which is applied directly to po.buff.
- partitioning/bmm/looper/bmm_looper.cpp: dst batch stride fallback
  (M * ldc).
- Addressed example issues

Clears CodeQL cpp/integer-multiplication-cast-to-long alerts 84, 85,
108 and 109.

Alerts 106 and 107 (cpp/suspicious-add-sizeof in
group_matmul/custom_kernel/dispatch.cpp) are intentionally not
addressed: the base pointer there is std::byte *, so scaling by
sizeof(bfloat16_t) is the correct element-to-byte conversion.

+++ zendnnl/gtests/group_matmul/test_basic.cpp
@@ -263,7 +263,7 @@ protected:
         const int num_slots = num_tokens * topk;
         std::vector<float> moe_weights(num_slots, 1.0f / topk);
         const size_t dst_elem_sz = zendnnl::common::size_of(dst_dt);
-        std::vector<char> moe_output(num_tokens * D * dst_elem_sz, 0);
+        std::vector<char> moe_output((size_t)num_tokens * D * dst_elem_sz, 0);
         std::vector<const void *> row_ptrs(num_slots);
 
         group_matmul_moe_postop_params moe {};
